moves add_to_db to functions
