--- docs/api/aim.digifeeds.add_to_db.rst
@@ -1,7 +1,0 @@
-aim.digifeeds.add\_to\_db module
-================================
-
-.. automodule:: aim.digifeeds.add_to_db
-   :members:
-   :undoc-members:
-   :show-inheritance:
