+++ docs/api/aim.digifeeds.rst
@@ -20,7 +20,6 @@ Submodules
 .. toctree::
    :maxdepth: 4
 
-   aim.digifeeds.add_to_db
    aim.digifeeds.alma_client
    aim.digifeeds.db_client
    aim.digifeeds.functions
